chore: remove unused GiftArticle code
ENTST-545
