--- components/x-gift-article/src/Buttons.scss
@@ -1,23 +1,0 @@
-@import '@financial-times/o-buttons/main';
-
-.x-gift-article__button {
-	@include oButtonsContent(
-		$opts: (
-			'type': 'primary',
-			'size': 'big'
-		)
-	);
-}
-
-.x-gift-article__buttons {
-	grid-area: buttons;
-	text-align: left;
-	white-space: nowrap;
-	margin-top: 12px;
-	display: flex;
-	align-items: end;
-}
-
-.x-gift-article-button--gap {
-	margin-right: 5px;
-}
