+++ components/x-gift-article/src/Footer.jsx
