--- components/x-gift-article/src/GiftArticle.scss
@@ -1,101 +1,0 @@
-@import './lib/variables';
-@import '@financial-times/o-normalise/main';
-@import '@financial-times/o-forms/main';
-@import '@financial-times/o-message/main';
-@import '@financial-times/o-typography/main';
-@import '@financial-times/o-labels/main';
-@import '@financial-times/o-colors/main';
-@import '@financial-times/o-icons/main';
-
-@import 'Buttons.scss';
-@import 'Loading.scss';
-@import 'Message.scss';
-@import 'MobileShareButtons.scss';
-@import 'HighlightSection.scss';
-@import 'HighlightRecipientMessage.scss';
-@import 'HighlightSavedMessage.scss';
-
-.x-gift-article {
-	@include oTypographySans;
-	strong {
-		font-weight: 600;
-	}
-}
-
-.x-gift-article__form {
-	max-width: none;
-	padding: 0;
-	margin: 0;
-
-	@include oForms(
-		$opts: (
-			'elements': (
-				'text',
-				'radio-round',
-				'checkbox'
-			),
-			'features': (
-				'inline',
-				'disabled'
-			)
-		)
-	);
-}
-
-.x-gift-article__radio_buttons {
-	margin-bottom: 12px;
-	@include oLabels(
-		$opts: (
-			'states': (
-				'content-premium'
-			)
-		)
-	);
-}
-
-.x-gift-article__enterprise-no-credits-icon {
-	@include oIconsContent($icon-name: 'warning', $color: oColorsByName('crimson'), $size: 16);
-	border-radius: 50%;
-	border: 2px solid oColorsByName('crimson');
-	margin: 0 0 -5px 5px;
-}
-
-@media only screen and (min-width: 600px) {
-	.x-gift-article__url-section {
-		display: grid;
-		grid-template-columns: auto min-content;
-		grid-template-rows: auto auto;
-		grid-template-areas:
-			'share-url buttons'
-			'message message';
-		grid-column-gap: 20px;
-	}
-}
-
-.x-gift-article__title {
-	@include oTypographyHeading(4);
-}
-
-.x-gift-article__url-input {
-	grid-area: share-url;
-	max-width: none;
-}
-
-.x-gift-article__copy-confirmation {
-	margin-top: 8px;
-}
-
-@include oMessage(
-	$opts: (
-		'types': (
-			'alert'
-		),
-		'states': (
-			'success'
-		)
-	)
-);
-
-.x-gift-article--visually-hidden {
-	@include oNormaliseVisuallyHidden();
-}
