--- components/x-gift-article/src/HighlightSavedMessage.scss
@@ -1,3 +1,0 @@
-.x-gift-article__highlights-saved-message {
-  margin-bottom: oSpacingByName('s4');
-}
