--- components/x-gift-article/src/Loading.jsx
@@ -1,7 +1,0 @@
-import { h } from '@financial-times/x-engine'
-
-export default () => (
-	<div className="x-gift-article-loading">
-		<div className="x-gift-article-loading__spinner" />
-	</div>
-)
