--- components/x-gift-article/src/Loading.scss
@@ -1,25 +1,0 @@
-@import '@financial-times/o-loading/main';
-
-.x-gift-article-loading {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	position: absolute;
-	top: 0;
-	right: 0;
-	bottom: 0;
-	left: 0;
-}
-
-.x-gift-article-loading__spinner {
-	@include oLoading(
-		$opts: (
-			'themes': (
-				'dark'
-			),
-			'sizes': (
-				'large'
-			)
-		)
-	);
-}
