+++ components/x-gift-article/src/ShareArticleDialog.scss
@@ -1,4 +1,4 @@
-@import '../lib/variables';
+@import './lib/variables';
 @import '@financial-times/o-normalise/main';
 @import '@financial-times/o-typography/main';
 @import '@financial-times/o-buttons/main';
