+++ components/x-gift-article/src/UrlSection.jsx
@@ -1,84 +1,59 @@
 import { h } from '@financial-times/x-engine'
+import CopyConfirmation from './CopyConfirmation'
 import { ShareType } from './lib/constants'
-import Url from './Url'
-import Message from './Message'
-import Buttons from './Buttons'
+import { SocialShareButtons } from './SocialShareButtons'
 
-export default ({
-	shareType,
-	isGiftUrlCreated,
-	isFreeArticle,
-	url,
-	urlType,
-	giftCredits,
-	monthlyAllowance,
-	nextRenewalDateText,
-	mailtoUrl,
-	redemptionLimit,
-	showCopyButton,
-	nativeShare,
-	invalidResponseFromApi,
-	actions,
-	enterpriseLimit,
-	enterpriseHasCredits,
-	enterpriseRequestAccess,
-	enterpriseFirstTimeUser,
-	includeHighlights
-}) => {
-	const hideUrlShareElements =
-		(giftCredits === 0 && shareType === ShareType.gift) ||
-		((enterpriseRequestAccess || !enterpriseHasCredits) && shareType === ShareType.enterprise)
-	const showUrlShareElements = !hideUrlShareElements
+export const UrlSection = (props) => {
+	const {
+		urlType,
+		url,
+		actions,
+		isArticleSharingUxUpdates,
+		shareType,
+		showCopyConfirmation,
+		enterpriseEnabled
+	} = props
+
+	const copyLinkHandler = (event) => {
+		switch (shareType) {
+			case ShareType.gift:
+				actions.copyGiftUrl(event)
+				break
+			case ShareType.enterprise:
+				actions.copyEnterpriseUrl(event)
+				break
+			case ShareType.nonGift:
+				actions.copyNonGiftUrl(event)
+				break
+			default:
+		}
+	}
 
 	return (
-		<div
-			className="js-gift-article__url-section x-gift-article__url-section"
-			data-section-id={shareType + 'Link'}
-			data-trackable={shareType + 'Link'}
-		>
-			{showUrlShareElements && (
-				<Url
-					{...{
-						shareType,
-						isGiftUrlCreated,
-						url,
-						urlType
-					}}
-				/>
-			)}
-
-			<Message
-				{...{
-					shareType,
-					isGiftUrlCreated,
-					isFreeArticle,
-					giftCredits,
-					monthlyAllowance,
-					nextRenewalDateText,
-					redemptionLimit,
-					invalidResponseFromApi,
-					enterpriseHasCredits,
-					enterpriseLimit,
-					enterpriseRequestAccess,
-					enterpriseFirstTimeUser
-				}}
-			/>
-
-			{showUrlShareElements && (
-				<Buttons
-					{...{
-						shareType,
-						isGiftUrlCreated,
-						mailtoUrl,
-						showCopyButton,
-						nativeShare,
-						actions,
-						giftCredits,
-						isFreeArticle,
-						includeHighlights
-					}}
-				/>
-			)}
+		<div>
+			<div
+				className="o-forms-input o-forms-input--text o-forms-input--suffix js-gift-article__url-section"
+				data-section-id={shareType + 'Link'}
+				data-trackable={shareType + 'Link'}
+			>
+				<input id="share-link" type="text" name={urlType} value={url} readOnly />
+				<button
+					className={`o-buttons o-buttons--big o-buttons--primary ${
+						enterpriseEnabled ? 'o-buttons--professional' : ''
+					}`}
+					aria-label="Copy link of the gift article to your clipboard"
+					onClick={copyLinkHandler}
+				>
+					Copy Link
+				</button>
+				{showCopyConfirmation && (
+					<CopyConfirmation
+						hideCopyConfirmation={actions.hideCopyConfirmation}
+						isArticleSharingUxUpdates={isArticleSharingUxUpdates}
+					/>
+				)}
+			</div>
+			<SocialShareButtons {...props} />
 		</div>
 	)
 }
