+++ components/x-gift-article/src/main.scss
@@ -1,2 +1,1 @@
-@import "GiftArticle.scss";
-@import "./v2/ShareArticleDialog.scss";
+@import "ShareArticleDialog";
